Wrote tests for DAO classes

+++ server/src/main/java/dataaccess/MemoryAuthDAO.java
@@ -8,7 +8,11 @@ import java.util.Map;
 
 public class MemoryAuthDAO implements AuthDAO {
 
-    private Map<String, AuthData> database = new HashMap<>();
+    private Map<String, AuthData> database;
+
+    public MemoryAuthDAO() {
+        database = new HashMap<>();
+    }
 
     public void clearAuth() {
         database.clear();
@@ -19,10 +23,16 @@ public class MemoryAuthDAO implements AuthDAO {
         database.put(newAuthToken, newAuthData);
         return newAuthData;
     }
-    public AuthData getAuth(String authToken) {
+    public AuthData getAuth(String authToken) throws DataAccessException {
+        if (!database.containsKey(authToken)) {
+            throw new DataAccessException("Invalid authToken");
+        }
         return database.get(authToken);
     }
-    public void deleteAuth(String authToken) {
+    public void deleteAuth(String authToken) throws DataAccessException {
+        if (!database.containsKey(authToken)) {
+            throw new DataAccessException("Cannot delete non-existent authToken");
+        }
         database.remove(authToken);
     }
 }
